refactor: Move manifests, configs, and references around
