+++ .pre-commit-config.yaml
@@ -4,60 +4,54 @@ default_language_version:
   python: python3.11
 
 repos:
-- repo: https://github.com/astral-sh/ruff-pre-commit
-  # Ruff version.
-  rev: v0.0.278
-  hooks:
-    - id: ruff
-      args: [ --fix, --exit-non-zero-on-fix ]
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    # Ruff version.
+    rev: v0.0.278
+    hooks:
+      - id: ruff
+        args: [--fix, --exit-non-zero-on-fix]
 
-- repo: local
-  hooks:
-  - id: isort
-    stages: [commit,push]
-    name: isort
-    entry: poetry run isort
-    language: system
-    types: [python]
-  - id: black
-    stages: [commit,push]
-    name: black
-    entry: poetry run black -S .
-    language: system
-    types: [python]
-  - id: mypy
-    stages: [commit,push]
-    name: mypy
-    entry: poetry run mypy --ignore-missing-imports
-    language: system
-    types: [python]
+  - repo: local
+    hooks:
+      - id: ruff
+        stages: [commit, push]
+        name: ruff_format
+        entry: poetry run ruff format
+        language: system
+        types: [python]
+      - id: mypy
+        stages: [commit, push]
+        name: mypy
+        entry: poetry run mypy --ignore-missing-imports
+        language: system
+        types: [python]
 
-- repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v2.1.0
-  hooks:
-  - id: trailing-whitespace
-    stages: [commit,push]
-  - id: check-added-large-files
-  - id: check-ast
-    stages: [commit,push]
-  - id: check-case-conflict
-  - id: check-byte-order-marker
-  - id: check-executables-have-shebangs
-  - id: check-docstring-first
-    stages: [commit,push]
-  - id: check-json
-  - id: check-merge-conflict
-    stages: [commit,push]
-  - id: check-symlinks
-  - id: check-vcs-permalinks
-  - id: check-xml
-  - id: check-yaml
-  - id: debug-statements
-  - id: detect-private-key
-  # - id: flake8
-    # stages: [commit,push]
-  - id: forbid-new-submodules
-  - id: no-commit-to-branch
-    stages: [commit,push]
-    args:
-    - --branch=main
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v2.1.0
+    hooks:
+      - id: trailing-whitespace
+        stages: [commit, push]
+      - id: check-added-large-files
+      - id: check-ast
+        stages: [commit, push]
+      - id: check-case-conflict
+      - id: check-byte-order-marker
+      - id: check-executables-have-shebangs
+      - id: check-docstring-first
+        stages: [commit, push]
+      - id: check-json
+      - id: check-merge-conflict
+        stages: [commit, push]
+      - id: check-symlinks
+      - id: check-vcs-permalinks
+      - id: check-xml
+      - id: check-yaml
+      - id: debug-statements
+      - id: detect-private-key
+      # - id: flake8
+      # stages: [commit,push]
+      - id: forbid-new-submodules
+      - id: no-commit-to-branch
+        stages: [commit, push]
+        args:
+          - --branch=main
