+++ pyproject.toml
@@ -19,12 +19,12 @@ boto3 = "^1.28.84"
 azure-storage-blob = "^12.19.0"
 azure-identity = "^1.15.0"
 types-pyyaml = "^6.0.12.12"
+types-networkx = "^3.2.1.20240313"
 
 [tool.poetry.group.dev.dependencies]
-ruff = "^0.0.278"
+ruff = "^0.3.0"
 pytest = "^7.4.0"
 isort = "^5.12.0"
-black = "^23.7.0"
 dbt-duckdb = "^1.6.0"
 duckdb = ">=0.8.0"
 pre-commit = "^3.6.0"
